+++ parser/clj_kondo/impl/rewrite_clj/node/quote.clj
@@ -62,7 +62,7 @@
   [children]
   (if (sequential? children)
     (->node
-      :unquote "~" 'unquote
+      :unquote "~" 'clojure.core/unquote
       children)
     (recur [children])))
 
@@ -72,6 +72,6 @@
   [children]
   (if (sequential? children)
     (->node
-      :unquote-splicing "~@" 'unquote-splicing
+      :unquote-splicing "~@" 'clojure.core/unquote-splicing
       children)
     (recur [children])))