+++ src/clj_kondo/impl/analyzer.clj
@@ -101,7 +101,6 @@
         (findings/reg-finding!
          ctx
          {:message "Keys in :or should be simple symbols."
-          :level :error
           :row (:row m)
           :col (:col m)
           :end-row (:end-row m)
@@ -193,7 +192,6 @@
                     ctx
                     (node->line (:filename ctx)
                                 expr
-                                :error
                                 :syntax
                                 (str "unsupported binding form " sym)))))))
            ;; keyword
@@ -222,7 +220,6 @@
                   ctx
                   (node->line (:filename ctx)
                               expr
-                              :error
                               :syntax
                               (str "unsupported binding form " k))))))
            :else
@@ -230,7 +227,6 @@
             ctx
             (node->line (:filename ctx)
                         expr
-                        :error
                         :syntax
                         (str "unsupported binding form " expr))))
          :vector (let [children (:children expr)
@@ -302,7 +298,6 @@
           ctx
           (node->line (:filename ctx)
                       expr
-                      :error
                       :syntax
                       (str "unsupported binding form " expr))))))))
 
@@ -350,7 +345,6 @@
         (findings/reg-finding! ctx
                                (node->line (:filename ctx)
                                            body
-                                           :warning
                                            :syntax
                                            "Function arguments should be wrapped in vector."))
         (let [arg-bindings (extract-bindings ctx arg-vec (:fn-body body) {:fn-args? true})
@@ -368,12 +362,12 @@
     (findings/reg-finding! ctx
                            (node->line (:filename ctx)
                                        body
-                                       :warning
                                        :syntax
                                        "Invalid function body."))))
 
 (defn analyze-pre-post-map [ctx expr]
-  (let [children (:children expr)]
+  (let [children (:children expr)
+        ctx (update ctx :callstack cons [:pre-post-map])]
     (key-linter/lint-map-keys ctx expr)
     (mapcat (fn [[key-expr value-expr]]
               (let [analyzed-key (analyze-expression** ctx key-expr)
@@ -414,7 +408,6 @@
                 (findings/reg-finding! ctx
                                        (node->line (:filename ctx)
                                                    first-child
-                                                   :warning
                                                    :misplaced-docstring
                                                    "Misplaced docstring."))))
         [parsed return-tag]
@@ -486,7 +479,6 @@
             (findings/reg-finding! ctx
                                    (node->line (:filename ctx)
                                                expr
-                                               :warning
                                                :syntax
                                                "Invalid function body.")))
         ;; var is known when making recursive call
@@ -622,7 +614,7 @@
     (when (odd? num-children)
       (findings/reg-finding!
        ctx
-       (node->line (:filename ctx) bv :error :syntax
+       (node->line (:filename ctx) bv :syntax
                    (format "%s binding vector requires even number of forms" form-name))))))
 
 (defn assert-vector [ctx call expr]
@@ -631,7 +623,7 @@
       (if-not vec?
         (do (findings/reg-finding!
              ctx
-             (node->line (:filename ctx) expr :error :syntax
+             (node->line (:filename ctx) expr :syntax
                          ;; cf. error in clojure
                          (format "%s requires a vector for its binding" call)))
             nil)
@@ -658,7 +650,7 @@
                    (and valid-bv-node (empty? (:children valid-bv-node)))))
       (findings/reg-finding!
        ctx
-       (node->line filename expr :warning :redundant-let "Redundant let expression.")))
+       (node->line filename expr :redundant-let "Redundant let expression.")))
     (when bv-node
       (let [{analyzed-bindings :bindings
              arities :arities
@@ -709,7 +701,7 @@
     (when redundant?
       (findings/reg-finding!
        ctx
-       (node->line filename expr :warning :redundant-do "redundant do"))))
+       (node->line filename expr :redundant-do "redundant do"))))
   (analyze-children ctx (next (:children expr))))
 
 (defn lint-two-forms-binding-vector! [ctx form-name expr]
@@ -717,7 +709,7 @@
     (when (not= 2 num-children)
       (findings/reg-finding!
        ctx
-       (node->line (:filename ctx) expr :error :syntax (format "%s binding vector requires exactly 2 forms" form-name))))))
+       (node->line (:filename ctx) expr :syntax (format "%s binding vector requires exactly 2 forms" form-name))))))
 
 (defn analyze-conditional-let [ctx call expr]
   (let [children (next (:children expr))
@@ -828,7 +820,6 @@
            (node->line
             filename
             expr
-            :warning
             :unexpected-recur "unexpected recur"))
           (not= expected-arity arg-count)
           (findings/reg-finding!
@@ -836,7 +827,6 @@
            (node->line
             filename
             expr
-            :error
             :invalid-arity
             (format "recur argument count mismatch (expected %d, got %d)" expected-arity arg-count)))
           :else nil))))
@@ -876,12 +866,28 @@
 
 (declare analyze-defmethod)
 
+(defn current-namespace-var-name [ctx var-name-node var-name]
+  (if (qualified-symbol? var-name)
+    (let [ns (:ns ctx)
+          alias (symbol (namespace var-name))]
+      (if (= (:name ns) (get (:qualify-ns ns) alias))
+        (symbol (name var-name))
+        (do (findings/reg-finding!
+             ctx
+             (node->line (:filename ctx) var-name-node
+                         :syntax
+                         (str "Invalid var name: " var-name)))
+            nil)))
+    (some-> var-name
+            (with-meta (meta var-name-node)))))
+
 (defn analyze-def [ctx expr defined-by]
   ;; (def foo ?docstring ?init)
   (let [children (next (:children expr))
         var-name-node (->> children first (meta/lift-meta-content2 ctx))
         metadata (meta var-name-node)
         var-name (:value var-name-node)
+        var-name (current-namespace-var-name ctx var-name-node var-name)
         docstring (when (> (count children) 2)
                     (string-from-token (second children)))]
     (when var-name
@@ -922,7 +928,7 @@
         fn-meta (meta fn-name)]
     (when-let [k (types/keyword binding)]
       (when-not (types/match? k :ifn)
-        (findings/reg-finding! ctx (node->line (:filename ctx) expr :error
+        (findings/reg-finding! ctx (node->line (:filename ctx) expr
                                                :type-mismatch
                                                (format "%s cannot be called as a function."
                                                        (str/capitalize (types/label k)))))))
@@ -943,7 +949,7 @@
             (let [arg-count (count (rest children))]
               (when-not (arity-match? fixed-arities varargs-min-arity arg-count)
                 (findings/reg-finding! ctx
-                                       (node->line filename expr :error
+                                       (node->line filename expr
                                                    :invalid-arity
                                                    (linters/arity-error nil fn-name arg-count fixed-arities varargs-min-arity)))))))
         (analyze-children (update ctx :callstack conj [nil fn-name]) (rest children))))))
@@ -952,18 +958,27 @@
   (when (:in-def ctx)
     (findings/reg-finding!
      ctx
-     (node->line (:filename ctx) expr :warning :inline-def "inline def"))))
+     (node->line (:filename ctx) expr :inline-def "inline def"))))
 
 (defn analyze-declare [ctx expr defined-by]
   (let [ns-name (-> ctx :ns :name)
-        var-name-nodes (next (:children expr))]
-    (doseq [var-name-node var-name-nodes]
-      (namespace/reg-var! ctx ns-name
-                          (->> var-name-node (meta/lift-meta-content2 ctx) :value)
-                          expr
-                          (assoc (meta expr)
-                                 :declared true
-                                 :defined-by defined-by)))))
+        var-name-nodes (next (:children expr))
+        var-names (keep (fn [var-name-node]
+                          (let [var-sym (->> var-name-node (meta/lift-meta-content2 ctx) :value)]
+                            (current-namespace-var-name ctx var-name-node var-sym)))
+                        var-name-nodes)]
+    (doseq [var-name var-names]
+      (let [var-name-meta (meta var-name)]
+        (namespace/reg-var! ctx ns-name
+                            var-name
+                            expr
+                            (assoc (meta expr)
+                                   :name-row (:row var-name-meta)
+                                   :name-col (:col var-name-meta)
+                                   :name-end-row (:end-row var-name-meta)
+                                   :name-end-col (:end-col var-name-meta)
+                                   :declared true
+                                   :defined-by defined-by))))))
 
 (defn analyze-catch [ctx expr]
   (let [ctx (update ctx :callstack conj [nil 'catch])
@@ -1012,7 +1027,6 @@
              (node->line
               (:filename ctx)
               expr
-              :warning
               :missing-clause-in-try
               "Missing catch or finally in try")))
           analyzed)))))
@@ -1151,7 +1165,7 @@
       (findings/reg-finding!
        ctx
        (node->line (:filename ctx) not-expr
-                   :warning :not-empty?
+                   :not-empty?
                    "use the idiom (seq x) rather than (not (empty? x))")))
     (analyze-children ctx (rest (:children expr)) false)))
 
@@ -1180,7 +1194,7 @@
       (findings/reg-finding!
        ctx
        (node->line (:filename ctx) expr
-                   :warning linter
+                   linter
                    msg)))
     (analyze-children ctx args false)))
 
@@ -1251,7 +1265,6 @@
        (node->line
         (:filename ctx)
         expr
-        :warning
         :missing-body-in-when
         "Missing body in when"))
       (analyze-children ctx body false))))
@@ -1274,13 +1287,13 @@
                       (findings/reg-finding!
                        ctx
                        (node->line (:filename ctx) (last children)
-                                   :warning :type-mismatch
+                                   :type-mismatch
                                    "String match arg requires string replacement arg.")))
             :char (when (not (identical? matcher-type :char))
                     (findings/reg-finding!
                      ctx
                      (node->line (:filename ctx) (last children)
-                                 :warning :type-mismatch
+                                 :type-mismatch
                                  "Char match arg requires char replacement arg.")))
             :regex (when (not (or (identical? matcher-type :string)
                                   ;; we could allow :ifn here, but keywords are
@@ -1291,7 +1304,7 @@
                      (findings/reg-finding!
                       ctx
                       (node->line (:filename ctx) (last children)
-                                  :warning :type-mismatch
+                                  :type-mismatch
                                   "Regex match arg requires string or function replacement arg.")))
             nil))))))
 
@@ -1326,7 +1339,7 @@
         arg-count (count args)]
     (when-not (= percent-count
                  arg-count)
-      (findings/reg-finding! ctx (node->line (:filename ctx) format-str-node :error :format
+      (findings/reg-finding! ctx (node->line (:filename ctx) format-str-node :format
                                              (format "Format string expects %s arguments instead of %s."
                                                      percent-count arg-count))))))
 
@@ -1425,7 +1438,7 @@
                             fn-name (or fsym (:name fst-ana))]
                         (findings/reg-finding!
                          ctx
-                         (node->line filename f :error
+                         (node->line filename f
                                      :invalid-arity
                                      (linters/arity-error nil fn-name arg-count fixed-arities varargs-min-arity))))))))))
     (concat fana
@@ -1477,7 +1490,7 @@
                 ;; See #1170, we deliberaly use resolved and not resolved-as
                 ;; Users can get :lint-as like behavior for hooks by configuring
                 ;; multiple fns to target the same hook code
-                hook-fn (hooks/hook-fn ctx config :analyze-call resolved-namespace resolved-name)
+                hook-fn (hooks/hook-fn ctx config resolved-namespace resolved-name)
                 transformed (when hook-fn
                               ;;;; Expand macro using user-provided function
                               (let [filename (:filename ctx)]
@@ -1487,8 +1500,10 @@
                                                    :cljc (identical? :cljc base-lang)
                                                    :lang lang
                                                    :filename filename
-                                                   :config config})
+                                                   :config config
+                                                   :ns ns-name})
                                          (catch Exception e
+                                           ;;TODOAQUI
                                            (findings/reg-finding!
                                             ctx
                                             (merge
@@ -1503,6 +1518,7 @@
             (if-let [expanded (and transformed
                                    (:node transformed))]
               (do ;;;; This registers the macro call, so we still get arity linting
+                ;; (prn :expanded expanded)
                 (namespace/reg-var-usage! ctx ns-name {:type :call
                                                        :resolved-ns resolved-namespace
                                                        :ns ns-name
@@ -1827,7 +1843,7 @@
         (when (or (zero? arg-count)
                   (> arg-count 2))
           (findings/reg-finding! ctx
-                                 (node->line (:filename ctx) expr :error :invalid-arity
+                                 (node->line (:filename ctx) expr :invalid-arity
                                              (format "keyword :%s is called with %s args but expects 1 or 2"
                                                      kw-str
                                                      arg-count))))))))
@@ -1840,7 +1856,7 @@
       (when-not (config/skip? config :invalid-arity callstack)
         (findings/reg-finding!
          ctx
-         (node->line (:filename ctx) expr :error :invalid-arity
+         (node->line (:filename ctx) expr :invalid-arity
                      (format "map is called with %s args but expects 1 or 2"
                              arg-count)))))))
 
@@ -1851,7 +1867,7 @@
       (when-not (config/skip? config :invalid-arity callstack)
         (findings/reg-finding!
          ctx
-         (node->line (:filename ctx) expr :error :invalid-arity
+         (node->line (:filename ctx) expr :invalid-arity
                      (str "Vector can only be called with 1 arg but was called with: "
                           arg-count)))))))
 
@@ -1863,7 +1879,7 @@
       (when-not (config/skip? config :invalid-arity callstack)
         (findings/reg-finding!
          ctx
-         (node->line (:filename ctx) expr :error :invalid-arity
+         (node->line (:filename ctx) expr :invalid-arity
                      (format "symbol is called with %s args but expects 1 or 2"
                              arg-count)))))))
 
@@ -1874,7 +1890,7 @@
     (when-not (config/skip? config :not-a-function callstack)
       (findings/reg-finding!
        ctx
-       (node->line (:filename ctx) expr :error :not-a-function (str "a " typ " is not a function"))))))
+       (node->line (:filename ctx) expr :not-a-function (str "a " typ " is not a function"))))))
 
 (defn analyze-reader-macro [ctx expr]
   (analyze-children ctx (rest (:children expr))))