+++ src/clj_kondo/impl/hooks.clj
@@ -1,11 +1,15 @@
 (ns clj-kondo.impl.hooks
   {:no-doc true}
   (:require [clj-kondo.impl.findings :as findings]
+            [clj-kondo.impl.metadata :as meta]
+            [clj-kondo.impl.rewrite-clj.node :as node]
             [clj-kondo.impl.utils :as utils :refer [assoc-some vector-node list-node
                                                     sexpr token-node keyword-node
                                                     string-node map-node *ctx*]]
             [clojure.java.io :as io]
-            [sci.core :as sci]))
+            [clojure.walk :as walk]
+            [sci.core :as sci])
+  (:refer-clojure :exclude [macroexpand]))
 
 (set! *warn-on-reflection* true)
 
@@ -58,6 +62,32 @@
 (defn mark-generate [node]
   (assoc node :clj-kondo.impl/generated true))
 
+(defn coerce [s-expr]
+  (node/coerce s-expr))
+
+(defn annotate [node meta]
+  (walk/postwalk (fn [node]
+                   (if (map? node)
+                     (-> node
+                         (with-meta meta)
+                         mark-generate)
+                     node)) node))
+
+(defn -macroexpand [macro node bindings]
+  (let [call (sexpr node)
+        args (rest call)
+        res (apply macro call bindings args)
+        coerced (coerce res)
+        annotated (annotate coerced (meta node))
+        lifted (meta/lift-meta-content2 *ctx* annotated)]
+    ;;
+    lifted))
+
+#_(defmacro macroexpand [macro node]
+  `(clj-kondo.hooks-api/-macroexpand (deref (var ~macro)) ~node))
+
+(def ans (sci/create-ns 'clj-kondo.hooks-api nil))
+
 (def api-ns
   {'keyword-node (comp mark-generate keyword-node)
    'keyword-node? keyword-node?
@@ -73,7 +103,9 @@
    'list-node? list-node?
    'sexpr sexpr
    'reg-finding! reg-finding!
-   'reg-keyword! reg-keyword!})
+   'reg-keyword! reg-keyword!
+   'coerce coerce
+   })
 
 (def sci-ctx
   (sci/init {:namespaces {'clojure.core {'time (with-meta time* {:sci/macro true})}
@@ -103,26 +135,46 @@
           (swap! mem assoc args ret)
           ret)))))
 
+(def load-lock (Object.))
+
 (def hook-fn
   (let [delayed-cfg
-        (fn [ctx config k ns-sym var-sym]
+        (fn [ctx config ns-sym var-sym]
           (try (let [sym (symbol (str ns-sym)
-                                 (str var-sym))]
-                 (when-let [x (get-in config [:hooks k sym])]
-                   ;; we return a function of ctx, so we will never memoize on
-                   ;; ctx, which will hold on to all the linting state and
-                   ;; creates memory leaks for long lives processes (LSP /
-                   ;; VSCode), see #1036
-                   (sci/binding [sci/out *out*
-                                 sci/err *err*]
-                     (let [code (if (string? x)
-                                  (when (:allow-string-hooks ctx)
-                                    x)
-                                  ;; x is a function symbol
-                                  (let [ns (namespace x)]
-                                    (format "(require '%s)\n%s" ns x)))]
-                       (binding [*ctx* ctx]
-                         (sci/eval-string* sci-ctx code))))))
+                                 (str var-sym))
+                     hook-cfg (:hooks config)]
+                 (when hook-cfg
+                   (if-let [x (get-in hook-cfg [:analyze-call sym])]
+                     ;; we return a function of ctx, so we will never memoize on
+                     ;; ctx, which will hold on to all the linting state and
+                     ;; creates memory leaks for long lives processes (LSP /
+                     ;; VSCode), see #1036
+                     (sci/binding [sci/out *out*
+                                   sci/err *err*]
+                       (let [code (if (string? x)
+                                    (when (:allow-string-hooks ctx)
+                                      x)
+                                    ;; x is a function symbol
+                                    (let [ns (namespace x)]
+                                      (format "(require '%s)\n%s" ns x)))]
+                         (binding [*ctx* ctx]
+                           ;; require isn't thread safe in SCI
+                           (locking load-lock (sci/eval-string* sci-ctx code)))))
+                     (when-let [x (get-in hook-cfg [:macroexpand sym])]
+                       (sci/binding [sci/out *out*
+                                     sci/err *err*]
+                         (let [code (if (string? x)
+                                      (when (:allow-string-hooks ctx)
+                                        x)
+                                      ;; x is a function symbol
+                                      (let [ns (namespace x)]
+                                        (format "(require '%s)\n(deref (var %s))" ns x)))
+                               macro (binding [*ctx* ctx]
+                                       (locking load-lock
+                                         ;; require isn't thread safe in SCI
+                                         (sci/eval-string* sci-ctx code)))]
+                           (fn [{:keys [node]}]
+                             {:node (-macroexpand macro node (:bindings *ctx*))})))))))
                (catch Exception e
                  (binding [*out* *err*]
                    (println "WARNING: error while trying to read hook for"