+++ src/clj_kondo/impl/linters/deps_edn.clj
@@ -1,6 +1,7 @@
 (ns clj-kondo.impl.linters.deps-edn
   (:require [clj-kondo.impl.findings :as findings]
-            [clj-kondo.impl.utils :refer [sexpr node->line]]))
+            [clj-kondo.impl.utils :refer [sexpr node->line]]
+            [clojure.string :as str]))
 
 (set! *warn-on-reflection* true)
 
@@ -26,63 +27,94 @@
       vnodes)
     (take-nth 2 (rest (:children map-node)))))
 
-(defn lint-qualified-deps [ctx nodes]
-  (run! (fn [node]
-          (let [form (sexpr node)]
-            (when-not (or (qualified-symbol? form)
-                           ;; fix for namespaced maps
-                          (:namespace node))
-              (findings/reg-finding!
-               ctx
-               (node->line (:filename ctx)
-                           node
-                           :warning
-                           :deps.edn
-                           #_{:clj-kondo/ignore[:format]} ;; fixed on master
-                           (format "Libs must be qualified, change %s => %<s/%<s" form))))))
-        nodes))
+(defn lint-qualified-lib [ctx node]
+  (let [form (sexpr node)]
+    (when-not (or (qualified-symbol? form)
+                  ;; fix for namespaced maps
+                  (:namespace node))
+      (findings/reg-finding!
+       ctx
+       (node->line (:filename ctx)
+                   node
+                   :deps.edn
+                   #_{:clj-kondo/ignore[:format]} ;; fixed on master
+                   (format "Libs must be qualified, change %s => %<s/%<s" form)))
+      true)))
 
-(defn lint-dep-coordinates [ctx nodes]
-  (run! (fn [node]
-          (let [form (sexpr node)]
-            (if-not (map? form)
+(defn lint-qualified-deps [ctx nodes]
+  (run! #(lint-qualified-lib ctx %) nodes))
+
+(defn derive-git-url-from-lib [lib]
+  (when lib
+    (let [lib (str lib)]
+      (when (or
+             (str/starts-with? lib "io.")
+             (str/starts-with? lib "com."))
+        lib))))
+
+(defn lint-dep-coord [ctx lib node]
+  (let [form (sexpr node)
+        git-url (or (:git/url form)
+                    (derive-git-url-from-lib lib))]
+    (if-not (map? form)
+      (findings/reg-finding!
+       ctx
+       (node->line (:filename ctx)
+                   node
+                   :deps.edn
+                   (str "Expected map, found: " (.getName (class form)))))
+      (or (when-let [version (:mvn/version form)]
+            (when (or (= "RELEASE" version)
+                      (= "LATEST" version))
               (findings/reg-finding!
                ctx
                (node->line (:filename ctx)
                            node
-                           :warning
                            :deps.edn
-                           (str "Expected map, found: " (.getName (class form)))))
-              (or (when-let [version (:mvn/version form)]
-                    (when (or (= "RELEASE" version)
-                              (= "LATEST" version))
-                      (findings/reg-finding!
-                       ctx
-                       (node->line (:filename ctx)
-                                   node
-                                   :warning
-                                   :deps.edn
-                                   (str "Non-determistic version."))))
-                    true)
-                  (when (:git/url form)
-                    (when-not (:sha form)
-                      (findings/reg-finding!
-                       ctx
-                       (node->line (:filename ctx)
-                                   node
-                                   :warning
-                                   :deps.edn
-                                   (str "Missing required key :sha."))))
-                       true)
-                  (:local/root form)
-                  (findings/reg-finding!
-                   ctx
-                   (node->line (:filename ctx)
-                               node
-                               :warning
-                               :deps.edn
-                               (str "Missing required key: :mvn/version, :git/url or :local/root.")))))))
-        nodes))
+                           (str "Non-determistic version."))))
+            true)
+          (when (:git/url form)
+            (when (and (:git/sha form) (:sha form))
+              (findings/reg-finding!
+               ctx
+               (node->line (:filename ctx)
+                           node
+                           :deps.edn
+                           (str "Conflicting keys :git/sha and :sha."))))
+            (when (and (:git/tag form) (:tag form))
+              (findings/reg-finding!
+               ctx
+               (node->line (:filename ctx)
+                           node
+                           :deps.edn
+                           (str "Conflicting keys :git/tag and :tag."))))
+            (when-not (or (:git/sha form) (:sha form))
+              (findings/reg-finding!
+               ctx
+               (node->line (:filename ctx)
+                           node
+                           :deps.edn
+                           (str "Missing required key :git/sha."))))
+            true)
+          ;; no linting yet
+          (:local/root form)
+
+          ;; in this case git/url is inferred from lib
+          ;; see https://clojure.org/reference/deps_and_cli#_coord_attributes
+          (and git-url (:git/sha form))
+          (prn :git-url git-url)
+          ;; no condition met, generic warning
+          (findings/reg-finding!
+           ctx
+           (node->line (:filename ctx)
+                       node
+                       :deps.edn
+                       (str "Missing required key: :mvn/version, :git/url or :local/root.")))))))
+
+(defn lint-deps [ctx kvs]
+  (doseq [[lib coord] kvs]
+    (let [error? (lint-qualified-lib ctx lib)]
+      (lint-dep-coord ctx (when-not error? (sexpr lib)) coord))))
 
 (defn lint-alias-keys [ctx nodes]
   (run! (fn [node]
@@ -92,7 +124,6 @@
                ctx
                (node->line (:filename ctx)
                            node
-                           :warning
                            :deps.edn
                            (str "Prefer keyword for alias.")))
               (when (contains? #{:deps :extra-deps :jvm-opts} form)
@@ -100,7 +131,6 @@
                  ctx
                  (node->line (:filename ctx)
                              node
-                             :warning
                              :deps.edn
                              (str "Suspicious alias name: " (name form))))))))
         nodes))
@@ -115,7 +145,6 @@
                  ctx
                  (node->line (:filename ctx)
                              jvm-opts-node
-                             :warning
                              :deps.edn
                              (str "JVM opts should be seqable of strings.")))))))
         alias-nodes))
@@ -130,17 +159,19 @@
                  ctx
                  (node->line (:filename ctx)
                              repo-map-node
-                             :warning
                              :deps.edn
                              (str "Expected: map with :url."))))))
           repo-map-nodes)))
+
+(defn deps-map [raw-node]
+  (zipmap (-> raw-node key-nodes)
+          (-> raw-node val-nodes)))
 
 (defn lint-deps-edn [ctx expr]
   (try
     (let [deps-edn (sexpr-keys expr)
           deps (:deps deps-edn)
-          _ (lint-qualified-deps ctx (-> deps key-nodes))
-          _ (lint-dep-coordinates ctx (-> deps val-nodes))
+          _ (lint-deps ctx (deps-map deps))
           aliases (:aliases deps-edn)
           alias-keys (key-nodes aliases)
           _ (lint-alias-keys ctx alias-keys)
@@ -148,14 +179,14 @@
           alias-maps (map sexpr-keys alias-maps)
           _ (lint-aliases ctx alias-maps)
           extra-dep-maps (map :extra-deps alias-maps)
+          _ (run! #(lint-deps ctx (deps-map %)) extra-dep-maps)
           extra-dep-map-vals (mapcat val-nodes extra-dep-maps)
-          _ (lint-dep-coordinates ctx extra-dep-map-vals)
+          ;; _ (lint-dep-coordinates ctx extra-dep-map-vals)
           extra-dep-map-vals (map sexpr-keys extra-dep-map-vals)
           exclusions (map (comp :children :exclusions) extra-dep-map-vals)
-          _ (when-let [mvn-repos (:mvn/repos deps-edn)]
-              (lint-mvn-repos ctx mvn-repos))]
-      (run! #(lint-qualified-deps ctx (key-nodes %)) extra-dep-maps)
-      (run! #(lint-qualified-deps ctx %) exclusions))
+          _ (run! #(lint-qualified-deps ctx %) exclusions)]
+      (when-let [mvn-repos (:mvn/repos deps-edn)]
+        (lint-mvn-repos ctx mvn-repos)))
     ;; Due to ubiquitous use of sexpr, we're catching coercion errors here and let them slide.
     (catch Exception e
       (binding [*out* *err*]